FH-3018 Update README with multiple sync loop caveat
If only the `sync_frequency` value on a client dataset is changed
to a long timeout an issue can occur where the default 10 second
dataset timeout for a dataset on the server causes two sync loops
to be called.

This documents that the `syncFrequency` value of the dataset on
the server-side should also be updated to a value near that of the
client, otherwise multiple two sync loop issue documented here will
occur.