Fix the redis client creation when doing a `sync.connect()`
This change does 2 things to fix the below error on startup
```
Internal error: [Error: Redis connection to 127.0.0.1:[object Object]
failed - connect ECONNREFUSED 127.0.0.1]
```

First, the redis client is updated to the latest (2.6.5) from 0.8.2
(~5yrs old). The latest version is backwards compatible, despite the 2
major version bumps.
Second, the redis connection string is changed to a supported format
i.e. starts with `redis://`. This is the actual fix.