FH-3238 Configure mongodb-lock to remove expired locks
Currently, due to the way `mongodb-lock` works many records are being
created in MongoDB over a short period of time. This is causing issues
on `findAndModify` after some time.

This updates the branch of `mongodb-lock` to one that has an option
to remove records when they have expired. This may be published as
version `0.4.0` of `mongodb-lock`.