FH-3417 Ensure clients update when there is a collision JIRA: https://issues.jboss.org/browse/FH-3417
There are cases where a client will not update its records from a collision
when coming online from an offline state. This can result in a record not
being updated on a client until they update another record, which could take
a while. This updates the global hash of a dataset to also take into account
the amount of collisions for a dataset, when this is increased the global hash
of the dataset client will change, causing the client to perform a syncRecords
call and will then update the record locally.

This is done by adding a counter to the dataset client and incrementing it
each time there is a collision, this is then included when producing the
global hash.