Add error event listener for Redis client
Currently, when Redis goes down or is not available the Redis client
will emit an `error` event. In node, if an 'error' event does not have
a listener node will throw an Error and stop the process.

This change adds a listener that will produce a `console.warn` log,
this will prevent the app from crashing when the Redis client is down.