🐛 Ensure connect is finished before continuing to invoke sync
This change is to prevent a scenario where the application is
starting and a client is sending a sync request to it.
In that case, the `connect()` function may still be executing,
so sync is not ready to `start` yet. To prevent `start` being
executing in this case, an error will be sent to the sync client.
This will cause a 'crashed' notification and the client will retry
later.