⚡ Change the default sync worker interval & backoff settings
The default is to use an exponential backoff, ranging from 1ms to
1000ms.
These values are based on load tests.
The idea is to take full advantage of CPU when the sync queue is busy
and to allow the CPU to be used for other things when the sync queue is
not busy. Setting a max of 1000ms will ensure then should the server get
busy again, there won't be a significant delay in ramping up.