🐛 Fix timing stats for MongoDB Operations & API Process times
This fixes a regression introduced in
https://github.com/feedhenry/fh-sync/commit/eb543ed9dd1e6da6c2cf206428333f8aa56084ff.
The metrics key prefix will be applied to the key name at startup rather
than at metric send time. This will ensure the same key is used for both
metrics sending and retrieval.

A side-effect of this change is all other metrics will also be prefixed
with the app title, preventing cross app sending/retrieval of metrics if
more than 1 app is using the same metric backend (e.g. redis)